[Cleaning Round 1]" Client Initial Read Through
- Initial read through of the codebase
- Resharp the code files
- Cleanup Whitespaces
- Cleanup trailing line endings
- Cleanup console messages
- Cleanup Commented or dead code
- Update licensing text to new format

NOTE 1: Licensing - This updates the licensing text to the new format to point users viewing the source code to the Licenses directory where all Licenses will reside.  It is perfectly legal as it does not destroy the license.  Additionally it will allow for not needing to add new urls to the licensing text itself.

Note 2: Cleaning out the commented out code that is either dead or irrelevent will help reduce the space used on the physical hard drive and is less confusing.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>