[Cleaning Round 1]: Console Initial Read Through - Begin my initial read through of the code - Cleanup whitespaces - Cleanup unnecessary trailing line endings - Resharp files and correct formatting - Begin correcting console messages - Cleanup old and dead commented out code - Begin adding comments as to what each piece of code does for better history later on. - Organize using references properly - Update license to the new license format
NOTE 1: On license updates - The license stays the same but it will allow to put third party licenses in one directory while not adding the link to the site in the license text in the code files.  This is legal and still ensures compliance with all licenses while ensuring there is no confusion in the license text.

NOTE 2: Commented out code - While I agree that commented out code provides good historical reference, it is not necessary to keep it in the files and in fact increases the size of the space that file takes up on the physical disk drive.  By adding reference comments as to what things do will ensure that there is no confusion with commented out code that end users might play with and cause potential bugs.  Additionally this dead and commented out code will not apply to us once i do the WhiteCorification of the code.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>