Make water color configurable in the World Map
- Makes the color of the water that shows on the world map configurable (good bye cheesey blue water)
- Cleanup whitespaces
- Cleanup unnecessary trailing line endings
- Update configuration files
- Add necessary items in Util.cs for Nini configuration processes to work

NOTE: This gets rid of the cheesey bright blue water color on the world map in the viewers (and potentially website).  It also makes the water color correct and look like that in Secondlife.  However allows grid operators and owners to choose their own water colors if they wish in the configuration files.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>